% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Address2Basin.R
\name{Address2Basin}
\alias{Address2Basin}
\title{Thai River basin name}
\usage{
Address2Basin(address)
}
\arguments{
\item{address}{Character, a Vector of address}

\item{address[1]}{Character, Subdistrict}

\item{address[2]}{Character, District}

\item{address[3]}{Character, Province}
}
\value{
Main River basin name
}
\description{
Finding the main River basin name using Address
}
\examples{
Address2Basin(c("ลำปลาทิว","ลาดกระบัง","กรุงเทพมหานคร"))
Address2Basin(c("บางบอนเหนือ","บางบอน","กรุงเทพมหานคร"))
Address2Basin(c("บางบอน","บางบอน","กรุงเทพมหานคร"))

}
