% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Address2BasinTbl.R
\name{Address2BasinTbl}
\alias{Address2BasinTbl}
\title{Thai River basin name}
\usage{
Address2BasinTbl(address)
}
\arguments{
\item{address}{Character, data frame of address}

\item{address[, 1]}{Character, Subdistrict}

\item{address[, 2]}{Character, District}

\item{address[, 3]}{Character, Province}
}
\value{
Main River basin name, vector
}
\description{
Finding the main River basin name using Addresses in xlsx
}
\examples{
data(address)
address
Address2BasinTbl(address)

}
