% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MainBasin_Address}
\alias{MainBasin_Address}
\title{Data of Main River basin with address}
\format{
A data frame with 9092 rows and 6 variables:
\describe{
  \item{Basin_code}{code of main River basin}
  \item{Subdistrict_code}{code of subdistrict}
  \item{Basin_name}{name of main basin}
  \item{Subdistrict}{name of subdistrict}
  \item{District}{name of district}
  \item{Province}{name of province}
}
}
\source{
{ONWR}
}
\usage{
MainBasin_Address
}
\description{
Contains the address of 22 main Thai River basins
}
\examples{
data(MainBasin_Address)    # Lazy loading
}
\keyword{datasets}
