% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Address2BasinTbl.R
\name{Address2BasinTbl}
\alias{Address2BasinTbl}
\title{Thai River basin name}
\usage{
Address2BasinTbl(address_table)
}
\arguments{
\item{address_table}{Character, a Vector of address-table}

\item{address_table[, 1]}{Character, Subdistrict}

\item{address_table[, 2]}{Character, District}

\item{address_table[, 3]}{Character, Province}
}
\value{
Main River basin name, vector
}
\description{
Finding the main River basin name using Addresses in xlsx
}
\examples{
address_table <- system.file("extdata","Address_table.xlsx", package = "AllWaterThai")
addressv <- read.xlsx(address_table)
head(addressv)
Basin <- Address2BasinTbl(address_table)

}
